vq_gdos returns TRUE/FALSE, not -2
Return result of vq_gdos() was misinterpreted as being the one from vq_vgdos().
This only affected v_opnwk()/v_clswk calls with driver ids > 10.

+++ fvdi/include/globals.h
@@ -16,7 +16,7 @@ extern Virtual *non_fvdi_vwk;
 extern Workstation *non_fvdi_wk;
 extern Workstation *screen_wk;
 extern List *driver_list;
-extern long old_gdos;
+extern short old_gdos;
 
 extern short old_wk_handle;
